Fix package to follow correct Go version guidelines
The current package name
(github.com/terraform-providers/terraform-provider-tls) lacks a version
path and as such cannot be imported into other projects.

I currently depend on this provider in
github.com/vancluever/terraform-provider-acme, and this is effectively
keeping me from modernizing the provider due to a lack of ability to
upgrade testing dependencies.